[DSL/Model] All those methods you always wanted and more.  a.k.a. DB Review Final.
* Link reworking. 
* Indexes working.
* Pseudo-closures. Fun, fun.
* Propagated model changes (Server/Adapters).
* "Default" tags now in place (with patches).
* Accessor visibilities.
* MVC separation in DSL/mapping.vm/Pojo.vm .
* Property hierarchy refactor.
* Various other things throughout
  (like a little more testing...)
  (& cleaned up the queries...)


git-svn-id: file:///home/svn/omero/trunk@450 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/common/src/ome/util/builders/PojoOptions.java
@@ -141,7 +141,7 @@ public class PojoOptions
     	return this;
     }
     
-    public PojoOptions countsFor(Integer i){
+    public PojoOptions countsFor(Long i){
     	options.put(COUNTS,i);
     	return this;
     }
@@ -198,7 +198,6 @@ public class PojoOptions
      * Filtered by Group
      * ============================== */
 
-    /* FIXME : This is currently disabled. 
     public PojoOptions grp(Long i){
         options.put(GROUP,i);
         return this;
@@ -216,7 +215,6 @@ public class PojoOptions
     public Long getGroup(){
         return (Long) options.get(GROUP);
     }
-    */
     
     /* ==============================
      * Helpers
